refactor(client): modularize layout components and extract reusable logic
- Created modular components: `Header`, `Footer`, `DemoProduct`, `FooterNav`, `CategoryNav`, and `CouponCenter` for better separation of concerns.
- Moved `categories` to separate data file and `cart` logic to `useCart` hook.
- Removed Tailwind and PostCSS configs from `ui` package as they are no longer required.
