+++ apps/client/src/data/categories.ts
@@ -1,0 +1,35 @@
+import { Category } from '@/types/layout';
+
+// Categories for product navigation
+export const categories: Category[] = [
+	{
+		id: 'featured',
+		name: 'Featured',
+		icon: '⭐'
+	},
+	{
+		id: 'coffee',
+		name: 'Coffee',
+		icon: '☕'
+	},
+	{
+		id: 'tea',
+		name: 'Tea',
+		icon: '🍵'
+	},
+	{
+		id: 'smoothies',
+		name: 'Smoothies',
+		icon: '🥤'
+	},
+	{
+		id: 'juices',
+		name: 'Juices',
+		icon: '🧃'
+	},
+	{
+		id: 'specials',
+		name: 'Specials',
+		icon: '✨'
+	}
+];
