+++ apps/client/src/types/layout.ts
@@ -1,0 +1,5 @@
+export interface Category {
+	id: string;
+	name: string;
+	icon: string;
+}
